requirements.txt
Flask
flask-cors
youtube-transcript-api
python-dotenv
requests
marked